Fix: 修復 models.py 語法與完整更新為 IntegerField

+++ orders/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-# 必須引入 goods 裡的 Product，因為訂單明細需要關聯到商品
 from goods.models import Product
 
 class Order(models.Model):
@@ -7,10 +6,10 @@ class Order(models.Model):
     name = models.CharField(max_length=50, verbose_name="收件人姓名")
     phone = models.CharField(max_length=20, verbose_name="聯絡電話")
     address = models.TextField(verbose_name="收件地址")
-    total_amount = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="總金額")
+    # 這裡已經改成整數
+    total_amount = models.IntegerField(verbose_name="總金額")
     created_at = models.DateTimeField(auto_now_add=True, verbose_name="建立時間")
 
-    # 🔹 加上這段 class Meta，後台就會顯示漂亮的中文「訂單」
     class Meta:
         verbose_name = "訂單"
         verbose_name_plural = "訂單"
@@ -22,10 +21,10 @@ class OrderItem(models.Model):
     """訂單明細表"""
     order = models.ForeignKey(Order, related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.PROTECT, verbose_name="商品")
+    # 這裡已經改成整數
     price = models.IntegerField(verbose_name="購買單價")
     quantity = models.IntegerField(verbose_name="購買數量")
 
-    # 🔹 加上這段 class Meta，後台就會顯示漂亮的中文「訂單明細」
     class Meta:
         verbose_name = "訂單明細"
         verbose_name_plural = "訂單明細"
